+++ lib/src/brickhouse/brickhouse.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:layrz_models/layrz_models.dart';
 
+
 // Freezed
 part 'brickhouse.freezed.dart';
 part 'brickhouse.g.dart';